Enhancements
Added the ability to load expressions to watch from disk - default location is debug/watch_items.txt - with associated topic.
Added the ability to reset the watched item list to the default core expressions/values - with associated topic.
Recategorized debug topics to be a little easier to find/use, now that we have a few!